backend/pyproject.toml
[tool.poetry]
name = "learning-app-backend"
version = "0.1.0"
description = "Backend for the Dynamic Learning App"
authors = ["Antigravity <antigravity@google.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
openai = "^1.12.0"
pypandoc = "^1.12"
python-dotenv = "^1.0.0"
ddgs = "^9.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
